# API-WEB-SAUDE/src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "API de informações de Clínicas e Hospitais",
		"description": "Sobre a API de Clínicas e Hospitais",
		"contact": {
			"email": "ericspsantos@gmail.com"
		},
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "API WEB SAÚDE local"
		},
		{
			"url": "https://api-web-saude.vercel.app/",
			"description": "API WEB SAÚDE hospedada"
		}
	],
	"paths": {
...
		"/admin/novo-hospital": {
			"post": {
				"summary": "Cadastrar Hospital",
				"description": "Essa rota será responsável por cadastrar um hospital",
				"tags": ["Hospitais"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"description": "Dados do Hospital",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Hospital"
							},
							"examples": {
								"clinica": {
									"value": {
										"cep": "00000-000",
										"rua": "rua x",
										"numero": 123,
										"bairro": "bairro x",
										"cidade": "cidade x",
										"uf": "xx",
										"nome": "Hospital x",
										"imagem": "url_da_imagem",
										"horarioSemana": {
											"open": "08:00",
											"close": "17:00"
										},
										"Sabado": {
											"open": "09:00",
											"close": "13:00"
										},
										"longitude": -43.123456,
										"latitude": -22.987654,
										"especialidades": [
											"id_especialidade_1",
											"id_especialidade_2"
										],
										"instagram": "perfil_instagram",
										"email": "emailhospital.com",
										"whatsapp": "+55 11 1234-5678",
										"descricao": "Descrição do hospital"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Hospital cadastrado com Sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Hospital"
								}
							}
						}
					},
					"400": {
						"description": "Campos inválidos ou nome muito curto",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										},
										"Errors": {
											"type": "array",
											"items": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/admin/alterar-hospital/{id}": {
			"put": {
				"summary": "Alterar Hospital",
				"description": "Essa rota será responsável por alterar um hospital existente",
				"tags": ["Hospitais"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "ID do hospital a ser alterado"
					}
				],
				"requestBody": {
					"description": "Dados atualizados do Hospital",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Hospital"
							},
							"examples": {
								"clinica": {
									"value": {
										"cep": "00000-000",
										"rua": "rua x",
										"numero": 123,
										"bairro": "bairro x",
										"cidade": "cidade x",
										"uf": "xx",
										"nome": "Hospital x",
										"imagem": "url_da_imagem",
										"horarioSemana": {
											"open": "08:00",
											"close": "17:00"
										},
										"Sabado": {
											"open": "09:00",
											"close": "13:00"
										},
										"longitude": -43.123456,
										"latitude": -22.987654,
										"especialidades": [
											"id_especialidade_1",
											"id_especialidade_2"
										],
										"instagram": "perfil_instagram",
										"email": "emailhospital.com",
										"whatsapp": "+55 11 1234-5678",
										"descricao": "Descrição do hospital"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Hospital alterado com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Hospital"
								}
							}
						}
					},
					"400": {
						"description": "Campos inválidos ou nome muito curto",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										},
										"Errors": {
											"type": "array",
											"items": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Hospital não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/admin/deletar-hospital/{id}": {
			"delete": {
				"summary": "Deletar Hospital",
				"description": "Essa rota será responsável por deletar um hospital existente",
				"tags": ["Hospitais"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "ID do hospital a ser deletado"
					}
				],
				"responses": {
					"204": {
						"description": "Hospital deletado com sucesso"
					},
					"404": {
						"description": "Hospital não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/hospitais": {
			"get": {
				"summary": "Listar Hospitais",
				"description": "Essa rota será responsável por listar todos os hospitais",
				"tags": ["Hospitais"],
				"responses": {
					"200": {
						"description": "Lista de hospitais",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Hospital"
									}
								},
								"examples": {
									"hospitais": {
										"value": [
											{
												"nome": "Hospital x",
												"imagem": "url_da_imagem_A",
												"horarioSemana": {
													"open": "08:00",
													"close": "17:00"
												},
												"Sabado": {
													"open": "09:00",
													"close": "13:00"
												},
												"longitude": -43.123456,
												"latitude": -22.987654,
												"especialidades": [
													"id_especialidade_1",
													"id_especialidade_2"
												],
												"email": "email@hospitalA.com",
												"whatsapp": "+55 11 1234-5678",
												"descricao": "Descrição do Hospital A",
												"endereco": "id_endereco_A"
											},
											{
												"nome": "Hospital B",
												"imagem": "url_da_imagem_B",
												"horarioSemana": {
													"open": "09:00",
													"close": "18:00"
												},
												"Sabado": {
													"open": "10:00",
													"close": "14:00"
												},
												"longitude": -43.654321,
												"latitude": -22.123456,
												"especialidades": [
													"id_especialidade_2",
													"id_especialidade_3"
												],
												"email": "email@hospitalB.com",
												"whatsapp": "+55 11 9876-5432",
												"descricao": "Descrição do Hospital B",
												"endereco": {
													"value": {
														"cep": "00000-000",
														"rua": "rua x",
														"numero": 123,
														"bairro": "bairro x",
														"cidade": "cidade x",
														"uf": "xx"
													}
												}
											}
										]
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"Hospital": {
				"type": "object",
				"properties": {
					"nome": {
						"type": "string"
					},
					"imagem": {
						"type": "string"
					},
					"horarioSemana": {
						"type": "object",
						"properties": {
							"open": {
								"type": "string"
							},
							"close": {
								"type": "string"
							}
						}
					},
					"Sabado": {
						"type": "object",
						"properties": {
							"open": {
								"type": "string"
							},
							"close": {
								"type": "string"
							}
						}
					},
					"longitude": {
						"type": "number"
					},
					"latitude": {
						"type": "number"
					},
					"especialidades": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"email": {
						"type": "string"
					},
					"whatsapp": {
						"type": "string"
					},
					"cep": {
						"type": "string"
					},
					"rua": {
						"type": "string"
					},
					"numero": {
						"type": "string"
					},
					"bairro": {
						"type": "string"
					},
					"cidade": {
						"type": "string"
					},
					"uf": {
						"type": "string"
					}
				},
				"required": [
					"nome",
					"longitude",
					"latitude",
					"especialidades",
					"cep",
					"rua",
					"numero",
					"bairro",
					"cidade",
					"uf"
				]
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}